Move TeamAvatar to general components
