+++ packages/app/src/app/components/TeamAvatar/index.ts
@@ -1,0 +1,1 @@
+export { TeamAvatar } from './TeamAvatar';
