+++ packages/app/src/app/pages/common/Modals/TeamInviteModal/index.tsx
@@ -9,7 +9,7 @@ import history from 'app/utils/history';
 import { Element, Button, Text } from '@codesandbox/components';
 import css from '@styled-system/css';
 import { useMutation } from '@apollo/react-hooks';
-import { TeamAvatar } from 'app/pages/NewDashboard/Sidebar/TeamAvatar';
+import { TeamAvatar } from 'app/components/TeamAvatar';
 
 export const TeamInviteModal = () => {
   const {
